+++ torch/csrc/jit/frontend/tracer.cpp
@@ -670,7 +670,7 @@ void addInputs(Node* n, const char* name, at::Layout value) {
   detail::genericAddInput(n, static_cast<int64_t>(value));
 }
 void addInputs(Node* n, const char* name, at::ScalarType value) {
-  detail::genericAddInput(n, static_cast<int64_t>(value));
+  detail::genericAddInput(n, value);
 }
 void addInputs(Node* n, const char* name, at::MemoryFormat value) {
   detail::genericAddInput(n, static_cast<int64_t>(value));
